Change Main::new return value.
Currently the `Main::new` function returns a `Result<Self, Error>`
however this function doesn't need to return a Result. No operation
can fail during the `new` function. Also, there is no error type
inside this function.

The signature of `Main::new` was changed and the `?` were removed
from the callers to this function.