Don't switch to Linux containers on 2nd uninstall (#601)
If a user runs `Uninstall-SecurityDaemon -Force` and `C:\ProgramData\iotedge-moby` does not exist (e.g., because the user previously uninstalled), the script will call `docker -SwitchDaemon`, which configures Docker for Windows to use Linux containers. This means that uninstalling our moby can potentially mess with the customer's existing Docker for Windows installation.

The root cause was a function with multiple responsibilities. `Test-IsDockerRunning` checked to see whether a docker service is installed and running, but also switched the daemon to use Linux containers as a side effect. This behavior is only desireable in the install case, not uninstall.

I fixed the bug by sprouting a new function, `Set-ContainerOs`, which gets called immediately after `Test-IsDockerRunning` in the install codepath, and doesn't get called at all in the uninstall codepath.