Ensure socket files get deleted when iotedged starts on Windows (#605)
We already do this elsewhere in the code; we just needed to do it here as well.

Before iotedged fires up an endpoint, it deletes the UNIX domain socket file associated with that endpoint, if it exists. But in Windows, the logic to check whether the file exists always returns false because it's not doing the special steps required to check for the existence of a file _with a reparse point_, which is what a domain socket file is. For reasons I don't understand yet, this isn't a problem if a client has connected to the socket. But if the server binds the socket and then closes the connection, and no client ever connects, then an attempt to bind again will fail.

The fix is to call `fs::symlink_metadata` on Windows, rather than `std::path::Path::exists`. Also, we don't need the result of `fs::metadata` in Windows and it fails in this scenario (in fact, `Path::exists` is just a wrapper around `fs::metadata`), so only call it in Linux.

I tested this on Windows; thanks to @arsing for dev/test on Linux.