Update Windows installer to create a user-defined network for modules (#627)
Currently on Windows, we create all modules (edgeAgent, edgeHub, and any others) on the default 'nat' network, due to Docker for Windows constraints IIRC.

On Linux, we provide useful isolation by creating edgeAgent on the default 'bridge' network so it can talk to `iotedged`, but then all other modules are on their own, user-defined network, named 'azure-iot-edge'. They can freely communicate with one another without the need to expose ports to the host.

We want our network topology on Windows to align with what we do on Linux. These changes accomplish that.