Fix bug in Docker codegen'd client that prevents listing networks from working.
The result of `NetworkApi::network_list` returns a `Network`, which contains
an `Ipam`. The swagger spec defines `Ipam::options` to be an array of objects,
but in fact it is a single object.

Compare the actual struct at
https://github.com/docker/docker-ce/blob/18.06/components/engine/api/types/network/network.go#L12
with the swagger spec at
https://github.com/docker/docker-ce/blob/18.06/components/engine/api/swagger.yaml#L1476-L1478

This meant that the result of `NetworkApi::network_list` would fail to
deserialize. Since iotedged filters the networks to only contain the one
mentioned in config.yaml, and since the `options` value is usually `null`,
this has not been noticed until now. But if the user manually creates
the network, such as with

    docker network create azure-iot-edge --subnet 172.18.0.0/16 --gateway 172.18.0.1

then the `options` key becomes a non-null object, which triggers the bug.