Use ProgramData environment variable where needed
Use the ProgramData environment variable to address disk space issues on
IoT Core as C:\ProgramData does not point to the Data partition. This
also addresses cleanup on systems where Moby was installed to the wrong
location.

This commit intentionally avoid changes to the iotedge install root to
avoid addressing static values in the configuration file.