Use ProgramData environment variable for MobyDataRoot
ProgramData environment variable should be used on MobyDataRoot to avoid
installing moby container images into the OS partition on
multi-partition installations like IoT Core. This is not addressed for
IoT Edge itself as it would require deeper refactoring.

This commit also includes cleanup code for systems with moby installed
on the OS partition.