Don't stop updating modules until new one pulled
This change modifies the behavior of the Agent Core HealthRestartPlanner
to delay the stopping of containers which are to be replaced until the
replacement has been pulled.

The HealthRestartPlanner no longer creates a stop task for modules which
are to be updated initially, and instead defers the stopping of the
current module to the GroupCommand created by
DockerCommandFactory::UpdateAsync.  UpdateAsync first pulls the new
container, then stops and removes the current container.  The new
Container is the created, and started in the usual way.

fixes #641