Fix invalid DNS SAN entries (#795) (#802)
* Fix invalid DNS SAN entries (#795)

Changes here fix a situation where a edge device's host name that begins with number(s) [0-9] gets sanitized. For example host name "2019edgehost" is consumed as "edgehost". This has caused problems was observed when using VMs that begin with numbers since it appears to be permitted configuration contrary to RFC 1035.

The changes involve passing the configured host name as is into the SAN entry without any modifications. The module id DNS entry continues to be sanitized.

* Revert "Fix invalid DNS SAN entries (#795)"

This reverts commit a8148ccc46749252769ef85c488d9788599968a4.

* prepare_dns_san_entries was allowing characters that are not A-Za-z0-9 (#736)

* Fix invalid DNS SAN entries (#795)

Changes here fix a situation where a edge device's host name that begins with number(s) [0-9] gets sanitized. For example host name "2019edgehost" is consumed as "edgehost". This has caused problems was observed when using VMs that begin with numbers since it appears to be permitted configuration contrary to RFC 1035.

The changes involve passing the configured host name as is into the SAN entry without any modifications. The module id DNS entry continues to be sanitized.