Update Leaf Device E2E with scope capabilities (#792)
Changes deal with adding capabilities to allow leaf devices to be the in or out of the Edge device's scope. For SAS auth being in scope is optional but it is required for X.509 auth. Any X.509 auth based device not in scope will not be allowed to authenticate with the Edge.

Additional changes are mostly refactoring for better readability, usage and maintenance.

Note:

    * Currently C# client and service SDK revisions used are preview as they have not been formally released.

    * Warnings as errors was disabled for the LeafDevice project on account of an issue with the preview SDK. Once the SDK is patched and released, this flag can be re enabled.