Add support for DPS symmetric key attestation (#767)
Summary: The changes here implement the client code needed to register with DPS in order for DPS to resolve the Edge device to the configured IoT Hub. More about this capability in DPS can be learned here: https://docs.microsoft.com/en-us/azure/iot-dps/quick-create-simulated-device-symm-key

Overall the changes deal with parsing the config.yaml to obtain the device's symmetric key and an attestation test is performed with DPS using a SAS token. The changes are primarily relegated to the dps and provisioning crates and come along with tests.

**Note:** While the feature implementation is done via this PR, it should be noted that the DPS service has not enabled provisioning devices with Edge capabilities. As a result, best effort testing has taken place and it was observed that the device that DPS registration at the destination IoT Hub took place but it did not have the Edge capability enabled. A true E2E test can take place when DPS enables this in the service.