Authorize all processes in module (#858)
This change allows all processes in a module to access APIs on the workload endpoint with policy = "caller". It does so by replacing `docker ps` + `docker inspect` with a call to `docker top` for the given ID. It has the nice side effect of making fewer calls to docker.

It also improves calls to the management endpoint with policy = "module(id)". It replaces `docker ps` + `docker inspect` with a single call to `docker inspect`.

Changes include:
- The addition of `get` and `top` functions to the ModuleRuntime trait
- Corresponding implementations for DockerModuleRuntime, ModuleClient, KubeModuleRuntime, and Test implementations. Note that for ModuleClient and KubeModuleRuntime, and empty/panicing impl was sufficient since nobody calls the new functions through those structs.
- Changes to edgelet_core::Authorization::auth_caller and auth_module to use the new ModuleRuntime functions
- A new `ModuleTop` structure to hold the results of `top`
- Moved the existing `docker inspect` call + processing into its own function, so that both DockerModuleRuntime::get and DockerModule::runtime_state can use it
- Changes to edgelet_http::Authorization that rippled out from its use of edgelet_core::Authorization