Fix to X.509 test runner script (#875)
When invoking the iotedgequickstart application, without the -e option, it attempted to fetch the event hub connection string from the secret store and failed. It failed because the credentials to access the secret store were not provided to the quick start since this is not the test approach followed any more since pipelines provide access to key value secrets.