Ignore errors from apt-get update but not from apt-get install (#886)
`set -e; a && b` does not fail the script even if `a` exits with
non-zero code. So before this change, `apt-get update` or `apt-get install`
could fail for some reason but the script would assume they succeeded and
proceed.

We do want to allow `apt-get update` to fail without stopping the script,
since sometimes repositories can be unavailable and we don't want to block
checkins for that. However a `apt-get install` failure will cause failures
down the line from missing dependencies, so we still want to fail in that case.