Handle asynchronous errors from pulling container images.
If a container image pull fails asynchronously, such as if the iamge exists
but does not have the right architecture, the overall HTTP response has
an OK status code but the last status object in the response body is
an error.

Before this change, only the status code was used to determine the overall
success of the operation. With this change, the last status object is also
used.