Don't let `takeown` block on a user prompt.
`takeown` prompts the user to allow it to take ownership of a directory
if the user does not have the "List" permission for it. If this prompt
happens during the uninstaller, it blocks the uninstaller with no
indication that the user needs to press Y to continue.

With this change, the `/d Y` switch is used to answer Y to the prompt
automatically.

This is likely the reason for issues like #793