Windows packaging changes (#962)
* Enable Windows CAB packages (#807)

The installer script now installs a CAB package containing the IoT Edge daemon,
the Moby engine, and the Moby CLI.

Install-SecurityDaemon and Uninstall-SecurityDaemon have been renamed to
Install-IoTEdge and Uninstall-IoTEdge respectively. The original names have
been kept as aliases for backward-compatibility.

There are also three new commands - Deploy-IoTEdge, Initialize-IoTEdge and
Update-IoTEdge. Deploy-IoTEdge installs the CAB and Initialize-IoTEdge
generates the config.yaml. Install-IoTEdge just calls Deploy and Initialize
in sequence. Update-IoTEdge updates the CAB; there is no longer a need to
uninstall and reinstall to update the daemon.

The -DeleteMobyDataRoot and -DeleteConfig switches have been removed from\
Uninstall-IoTEdge since they only existed to maintain the container images
and config.yaml in update scenarios. Since Uninstall is no longer needed for
update scenarios, it will always delete the Moby data root and config.yaml

The Moby engine service is now always installed and started, even when
using Linux containers. Since it listens on a different named pipe, it will
not interfere with an existing Docker for Windows install.

* Fix homedir in config.yaml generated by Windows installer (#890)

* Cleanup logging (#889)

* Cleanup logging

* Don't let `takeown` block on a user prompt (#910)

`takeown` prompts the user to allow it to take ownership of a directory
if the user does not have the "List" permission for it. If this prompt
happens during the uninstaller, it blocks the uninstaller with no
indication that the user needs to press Y to continue.

With this change, the `/d y` switch is used to answer Y to the prompt
automatically.

This is likely the reason for issues like #793

* Fix of image auth regex (#913)

On a run of the E2E tests it was observed that the existing regex matched a comment containing the word 'auth' and the replace was consequently broken. This fix attempts to search for the auth under the agent section.

* Cleanup packaging to make it more suitable for the release job (#892)

* Cleanup packaging to make it more suitable for the release job

* Catch download failures properly

* Enable resigning of the package catalog