Use .NET Core installer task for linux arm64 images pipeline (#1043)
The arm64 pipeline in the images build is using a script to install .NET Core, but then recent changes to the build script are looking on the PATH for dotnet so it doesn't use what was installed. This change updates the pipeline to use the .NET Core Installer task instead of our custom script.