Revert edge_hsm_sas_auth_int integration test to use public TPM API
b5f281b6f6b7594bc4aa395c0088731e06400af3 changed this test to use the
`*tpm_store*` functions. But it fails to link on Windows since the functions
are not available to be linked to. Since it's an integration test,
it should be using the public libiothsm API anyway.

This change reverts the test to use the public libiothsm API again.