Fix synchronization for HSM lib services (#1156)
Ensure all accesses to the HSM are synchronized via a mutex owned by the higher level application (iotedged). Prior to this the TPM and Crypto could access the HSM concurrently as the ref counted mutex synchronized access to the handle and its use. With this change we ensure synchronous across the TPM, X509 and Crypto interfaces when accessing the HSM lib.