Disable access TPM in memory interface (#1245)
*  Disable access TPM in memory interface (#1059)

Remove ability to configure an in memory TPM vs a HW TPM via an environment variable and added tests. This has caused failures when users have tried to use the in memory implementation which is useful only for testing.

Essentially environment variable IOTEDGE_USE_TPM_DEVICE will be ignored by libiothsm and by default is built for use with a TPM device. To use the in memory implementation the library must be built using cmake flag USE_TEST_TPM_INTERFACE_IN_MEM.

* Revert edge_hsm_sas_auth_int integration test to use public TPM API (#1087)

b5f281b6f6b7594bc4aa395c0088731e06400af3 changed this test to use the
`*tpm_store*` functions. But it fails to link on Windows since the functions
are not available to be linked to. Since it's an integration test,
it should be using the public libiothsm API anyway.

This change reverts the test to use the public libiothsm API again.