Don't restart iotedged when config.yaml isn't ready (#1278)
On Linux (systemd) when iotedge is installed, the daemon is configured to always restart on failure. Its config.yaml file needs to be configured by the user after installation, so as soon as it starts the first time it will go into a fail/restart loop until it exceeds the start rate limit and enters a cooldown period (a failed state where manual intervention is required). This change maps a specific exit code for this condition (specifically, failure when the connection string in config.yaml hasn't been changed from its boilerplate value), and instructs systemd to prevent restart for this exit code.

This change will help with automated tests, which sometimes trip the rate limiting threshold and aren't equipped to do manual intervention.