Ensure PS exits with non-zero exit code if the Windows installer script hits an error
Before this change, the script wrote errors to stdout in red and then
exited successfully.

It's useful to exit with a non-zero code in this case for the benefit of
wrapper scripts that shell out to the installer. So the installer now uses
`throw` to ensure PS exits with 1 instead of 0.