Ensure PS exits with non-zero exit code if the Windows installer script hits an error (#1297)
Before this change, the script wrote some errors to stdout in red and then
exited successfully. For some other errors it used `throw`, which did result
in a non-zero exit code but using PS's unhandled exception message format.

It's useful to exit with a non-zero code for the benefit of wrapper scripts
that shell out to the installer. So with this change, the script consistently
writes errors to stdout in red *and* uses `throw` to exit with code 1.