Fix Windows to initialize logging before anything that can fail.
Before this change, the Windows console mode would exit silently if it failed
to parse the config file. Also, the Windows service mode would exit silently
if it could never start the service, such as if it could not connect to
the service controller.

With this change, these errors are logged.