Save logs for each end-to-end test (#1366)
For each end-to-end test that runs, save the daemon and module logs in the "current" directory (as defined by .NET Core). The format of a log filename is '{deviceid}-{test-name}-{module-name|"iotedged"}.log'. Note that we also create a logfile for iotedged and edgeAgent when test suite setup is complete.