[k8s] Use module original name as a module id for authentication (#1383)
In order to authenticate the edge module deployed on kubernetes runtime iotedged should be able to read the original module name and use it as a module identifier. The original module name is stored as an annotation for service account corresponding to edge module. During authentication, iotedged makes an additional call to kube api to get service account description and extracts the original module name from annotations.