Update packages used by E2E, longhaul and stress jobs.
ARM32 tests run on Raspbian 9, so they need to use the debian9 packages.
The ubuntu16.04 packages depend on libssl1.0.0 which does not exist in
Raspbian 9.

ARM64 tests run on Ubuntu 18.04, so they should use the ubuntu18.04 packages.
The ubuntu16.04 packages happen to work since ubuntu18.04 does have libssl1.0.0
but there is no reason to use the older distro's packages.