Preserve any extra properties in createOptions set by the user.
... even if they're not defined in Docker.DotNet's or Rust's types.

Since Docker.DotNet has not been updated in a long time, it is missing
many fields that have been added to Docker. While the user could set these in
their module's `createOptions`, they would be stripped out by Edge Agent
when it deserialized and re-serialized the `createOptions` in order to add
env vars, port bindings, etc.

While the Rust bindings are more recent, they still don't have everything.

Furthermore, in both C# and Rust, there is the possibility for errors in
the spec to manifest as errors in the generated code. This has led to
https://github.com/Azure/iotedge/issues/1116 and
https://github.com/Azure/iotedge/issues/1513 in the past.

With this change, Edge Agent uses its own custom types to deserialize
the `createOptions`. These types have strongly-typed fields for the properties
that Edge Agent modifies, and have a weakly-typed dictionary field annotated
with `JsonExtensionData` to hold all the rest.

Similarly, the Rust bindings related to create options have been changed to
only parse the few properties we use in our code, and leave the rest to be held
in a `#[serde(flatten)]`-annotated `HashMap<String, serde_json::Value>` field.

Given the very tiny number of fields Edge Agent and iotedged need to mutate,
this approach should be the easiest and need the least maintenance.

Fixes #1580