Add single quotes around the cron schedule (#1640)
We recently moved the scheduled build trigger for the service deployment build into the YAML definition. Previously is was declared in the UI. For some reason the trigger is not firing. A few different docs/forums suggest that the cron schedule needs to be quoted. So I'm trying that to see if we can get this trigger to work.