port to reqwest + client async/await improvements
Porting things over to reqest took a while, but it's definately worth it.
Redirect handling is really hard.

I've also reworked the way authentication caching works, switching
things over to a scope-based system, which results in a _lot_ more cache
hits.

While this is a great improvement, there are still a couple of hot-spots
in the code that could be improved:
- if two requests with disjoint scopes are performed, and each requires
a authentication handshake, one of them will block while the other
completes it's auth handshake. see the TODO in auth/mod.rs:60
- there's probably a way to return the raw Result<Blob> futures from the
dowload_image flow helper, but my Futures-fu isn't good enough to get
the types/lifetimes to work right.
- Right now, the scope header in the www-authenticate header is totally
ignored, and the scope cache uses the "best guess" scope given by the
various client API methods. This works okay, but it would be better to
actually parse that scope header and use it to sanity-check the expected
scopes. Also, if the server returns multiple scopes, we can pre-populate
the cache with those additional entries.