Add support for backup and restore of in-memory db stores. (#1759)
EdgeAgent and EdgeHub currently support an in-memory only store which can be enabled by setting the UsePersistentStorage configuration flag to false in the module’s configuration.

With this change, this in-memory store will be backed up on graceful shutdown and restored from on the next startup. This will help customers leverage the benefits of an in-memory only store (thus limiting disk writes) and not lose messages when a module is killed or the device restarts.

The feature can be enabled by setting the EnableNonPersistentStorageBackup environment variable to true.

A host disk-mapped folder can be mounted into the container to be used for storing backups and the path to it will have to be specified using the BackupFolder environment variable with its value equal to the path of the backup folder within the container (as specified in the volume mount).

If a backup folder is not explicitly specified, the backups will be created on the container's file system instead.